mod proveedor edt pt1

+++ php/eliminarProveedor.php
@@ -1,6 +1,8 @@
 <?php 
     include "./conexion.php";
-
-    $conexion->query("delete from proveedores where id_proveedor=".$_POST['id_proveedor']);
-    echo 'Proveedor eliminado';
+    if (isset($_POST['id_proveedor'])) {
+        $conexion->query("delete from proveedores where id_proveedor=".$_POST['id_proveedor']);
+        echo 'Proveedor eliminado';
+    }
+    header("Location: ../admin/proveedores.php");
 ?>
